refactor: 파일구조
