+++ src/pages/Members/Login/LoginFrom/style.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const LoginBody = styled.div`
+  height: 100vh;
+  position: relative;
+`;
+
+export const Heading = styled.h1`
+  font-size: 24px;
+  font-weight: bold;
+`;
