+++ src/pages/Members/Login/constants.js
@@ -1,0 +1,14 @@
+export const LOGIN_DATA = [
+  {
+    name: 'email',
+    text: '이메일 *',
+    type: 'text',
+    placeholder: 'ID@example.com',
+  },
+  {
+    name: 'password',
+    text: '비밀번호 *',
+    type: 'password',
+    placeholder: '비밀번호를 입력해주세요.',
+  },
+];
